refactor to make organization more clear
