+++ lib/src/main/java/badgerlog/networktables/PublisherUpdater.java
@@ -3,7 +3,7 @@ package badgerlog.networktables;
 import java.util.function.Supplier;
 
 /**
- * Automates periodic value publication to NetworkTables by linking a {@link Publisher} with a value {@link Supplier}.
+ * Automates periodic value publication to NetworkTables by linking a with a value {@link Supplier}.
  * Implements the {@link Updater} interface to enable scheduled updates, typically used in loops or timed tasks.
  *
  * @param <T> Type of data being published
